* Currently broken state, attempting to switch to trie. Commit so can push save-point to remote branch...

+++ redshift/tagger.pxd
@@ -28,7 +28,7 @@ cdef class Tagger:
     cdef dict _count_feats(self, Sentence* sent, TagState* p, TagState* g, int i)
 
 
-cdef TagState* extend_state(TagState* s, size_t clas, weight_t score, size_t cost,
+cdef TagState* extend_state(TagState* s, size_t clas, weight_t score, int cost,
                             Pool pool)
 
 cdef inline size_t get_p(TagState* s) nogil
